feat(630): refactor happy path function
