+++ tfs-functions/src/functions/process-queue-item.ts
@@ -1,0 +1,15 @@
+import { app, InvocationContext } from '@azure/functions';
+
+import { createGiftFacility } from '../utils/create-gift-facility';
+
+export async function processQueueItem(queueItem: unknown, context: InvocationContext): Promise<void> {
+  context.log('Gift requests queue function received item:', queueItem);
+  await createGiftFacility(queueItem, context);
+  context.log('Gift facility creation succeeded');
+}
+
+app.storageQueue('processQueueItem', {
+  queueName: 'gift-requests',
+  connection: 'AzureWebJobsStorage',
+  handler: processQueueItem,
+});
